refactor(file struct): create config folder for overall app settings const files and move files
