+++ lib/config/colors.dart
