+++ lib/screen/onboard/onboard_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:line_up_front_end/components/buttons.dart';
-import 'package:line_up_front_end/components/colors.dart';
-import 'package:line_up_front_end/screen/const/custom_font_weight.dart';
+import 'package:line_up_front_end/config/colors.dart';
+import 'package:line_up_front_end/config/custom_font_weight.dart';
 import 'package:line_up_front_end/screen/onboard/onboard_page.dart';
 import 'package:line_up_front_end/screen/root_screen.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
@@ -319,8 +319,7 @@ class LoginPage extends OnboardingPage {
             height: 54,
           ),
           onTap: () {
-            Navigator.pushReplacement(context,
-                MaterialPageRoute(builder: (context) => const RootScreen()));
+            Navigator.pushReplacementNamed(context, "home");
           },
         ),
         const SizedBox(
@@ -332,8 +331,7 @@ class LoginPage extends OnboardingPage {
             height: 54,
           ),
           onTap: () {
-            Navigator.pushReplacement(context,
-                MaterialPageRoute(builder: (context) => const RootScreen()));
+            Navigator.pushReplacementNamed(context, "home");
           },
         ),
         const SizedBox(
@@ -345,8 +343,7 @@ class LoginPage extends OnboardingPage {
             height: 54,
           ),
           onTap: () {
-            Navigator.pushReplacement(context,
-                MaterialPageRoute(builder: (context) => const RootScreen()));
+            Navigator.pushReplacementNamed(context, "home");
           },
         ),
         const SizedBox(
@@ -358,8 +355,7 @@ class LoginPage extends OnboardingPage {
             height: 54,
           ),
           onTap: () {
-            Navigator.pushReplacement(context,
-                MaterialPageRoute(builder: (context) => const RootScreen()));
+            Navigator.pushReplacementNamed(context, "home");
           },
         ),
       ],
@@ -381,35 +377,4 @@ class LoginPage extends OnboardingPage {
   }
 }
 
-class LoginScreen extends StatelessWidget {
-  const LoginScreen({super.key});
 
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Column(
-          children: [
-            const TextField(
-              decoration: InputDecoration(hintText: 'Username'),
-            ),
-            const TextField(
-              decoration: InputDecoration(hintText: 'Password'),
-              obscureText: true,
-            ),
-            ElevatedButton(
-              onPressed: () {
-                // 로그인 로직 처리 후 메인 화면으로 이동
-                Navigator.pushReplacement(
-                  context,
-                  MaterialPageRoute(builder: (context) => const RootScreen()),
-                );
-              },
-              child: const Text('Login'),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
